mysite/requirements.txt
Click==7.0
cycler==0.10.0
dj-database-url==0.5.0
Django==2.2.4
django-heroku==0.3.1
django-tinymce4-lite==1.7.5
Flask==1.1.1
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10.1
jsmin==2.2.2
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.0
numpy==1.16.4
psycopg2==2.8.3
PyMySQL==0.9.3
pyparsing==2.4.0
python-dateutil==2.8.0
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
Werkzeug==0.15.5
whitenoise==4.1.3
